introduce executer.Run and ExecutionData and DLCI
Notes: we reformulated the memory model s.t. everything is stored in
docker containers (including literals). This in order to bubble up
through the AST once we've treversed to its deepest depths, we only
need to return the DLCI which is a pointer to the docker
container (memory cell) which contains our data

Co-authored-by: Connor Walsh <c@polygon.pizza>
Co-authored-by: Mouse Reeve <mousereeve@riseup.net>

Git issue references:

Gamelan music currently playing: Gendhing Sumyar